7.29: Also checking if specific Format Clock MP3s are relevant.
As requested by Gerhard.

The player was originally only checking Format Clock MP3s for relevance if they
were in a directory specified in the playlist. ie the intent was to use it as
a simple form of advert scheduling. Explicitely-specified MP3s were not checked
for relevance. I assumed that in this case the user really wanted the MP3s to
play and that the relevance settings did not apply.

This commit corrects that. Explicitely-listed Format Clock MP3s are now also
checked for relevance, and skipped if they aren't. They are also skipped if
they aren't loaded in the schedule database.