get_playback_events_info() now anticipates the start of the next hour.
The player now sets the "number of milliseconds until interrupt due to hour
change" variable corerctly, instead of setting it to 0 immediately after the
hour change.

Previously, the player didn't know when an hour change was impending, and
could schedule something else to occur during the hour transition. This
could sometimes cause "time check" items (at the start of the hour) to
be missed (as a result of later logic which resets our "format clock delay"
factor variable at the start of each hour).