Added player::get_miss_promos_before_time()
Returns the time (no date component) before which ads will be marked as missed.

The exact logic depends on whether Format Clocks are enabled or not. When
Format Clocks are enabled we use the regular (miss ads before) time, or the
start of the hour, whichever is earlier. This is as requested by Gerhard.