Updated logic which determines during what period to query for adverts.
When Format Clocks are not enabled, the logic works as usual (between now minus
(mins to miss ads after) and now).

When Format Clocks are enabled, we now get ads between (either the start of
the current hour, or now minus (mins to miss ads after), whichever is earlier),
and the end of the current segment segment, or 10 minutes from now, whichever
is sooner. This is as per Gerhard's request.