7.38.1: No longer playing cached (non-forced time) promos when not allowed
When fetching a promo item, we now first check if the Format Clock segment
has changed. If it has, and the new segment does not allow promos (except for
"forced time" promos), and we have a list of promo items waiting to play,
then we clear the list (and log a message). In this case the player will
re-query the database for any "forced time" promos and we will play any that
are retrieved.

Also updated debian/changelog.