Log more info when jumping into songs
Log more info to the debug log when about to check if we should skip the
start of a song (because it starts quietly). Also log a line if we don't
jump.

Adding this because some songs on our USAVE testing box (31) aren't
being jumped into, even though the schedule database says they should
be.