Remove 'programming element cache'
In recent studio R&D, the caching seems to be causing problems.

Basically, they're causing too many segment and mp3 changes too fast, so we hit
the cache a lot instead of regenerating playlists.

This causes problems because some of the behavior they rely on doesn't work
well with caches (random sequence playlists stay in the same order as before).